Undo 64bit/vista changes after splitting them off to their own branch, at least until they're stable.
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk/pwsafe/pwsafe@1676 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ DDSupport.cpp
@@ -6,7 +6,6 @@
  * http://www.opensource.org/licenses/artistic-license.php
  */
 
-#include "stdafx.h"
 #include "DDSupport.h"
 #include "corelib/MyString.h"
 #include "corelib/util.h"
@@ -22,7 +21,7 @@ void CDDObject::DDSerialize(CSMemFile &outDDmemfile)
   m_item.SerializePlainText(v);
   size_t len = v.size();
   outDDmemfile.Write(&len, sizeof(len));
-  outDDmemfile.Write(&(*v.begin()), (UINT)v.size());
+  outDDmemfile.Write(&(*v.begin()), v.size());
   trashMemory(&(*v.begin()), v.size());
 }
 
@@ -32,7 +31,7 @@ void CDDObject::DDUnSerialize(CSMemFile &inDDmemfile)
   size_t lenlen = inDDmemfile.Read(&len, sizeof(len));
   ASSERT(lenlen == sizeof(len) && len != 0);
   vector<char> v(len);
-  size_t lenRead = inDDmemfile.Read(&(*v.begin()), (UINT)len);
+  size_t lenRead = inDDmemfile.Read(&(*v.begin()), len);
   ASSERT(lenRead == len);
   bool status = m_item.DeserializePlainText(v);
   ASSERT(status);
@@ -46,7 +45,7 @@ void CDDObList::DDSerialize(CSMemFile &outDDmemfile)
   POSITION Pos;
   CDDObject* pDDObject;
 
-  nCount = (int)GetCount();
+  nCount = GetCount();
 
   outDDmemfile.Write((void *)&nCount, sizeof(nCount));
   outDDmemfile.Write((void *)&m_bDragNode, sizeof(bool));
